Split src/tools.(h|cpp) in two: move zlib, Gumbo and libmagic-dependent functions to zimwriterfs/tools.(h|cpp)
