+++ .gitignore
@@ -27,7 +27,6 @@
 *.exe
 *.out
 *.app
-zimwriterfs
 
 # Temporary and test files
 *.kate-swp
