+++ src/zimwriterfs/meson.build
@@ -1,6 +1,7 @@
 
 sources = [
   'zimwriterfs.cpp',
+  'tools.cpp',
   '../tools.cpp',
   'article.cpp',
   'zimcreatorfs.cpp',
